fix(webapp): restore valid time-tracking server action exports
